upgade microservice structure + api gateway

+++ Marketplace-platform/src/main/java/docker_test/com/models/Cart.java
@@ -3,11 +3,11 @@ package docker_test.com.models;
 import java.time.LocalDateTime;
 
 public final class Cart {
-    private long cart_id;
-    private long user_id;
-    private long product_id;
-    private Long variant_id;
-    private int quantity;
+    private Integer cart_id;
+    private Integer user_id;
+    private Integer product_id;
+    private Integer variant_id;
+    private Integer quantity;
     private LocalDateTime added_at;
     private LocalDateTime updated_at;
 
@@ -17,7 +17,7 @@ public final class Cart {
         this.updated_at = LocalDateTime.now();
     }
 
-    public Cart(long cart_id, long user_id, long product_id, Long variant_id,
+    public Cart(Integer cart_id, Integer user_id, Integer product_id, Integer variant_id,
                 int quantity, LocalDateTime added_at, LocalDateTime updated_at) {
         this.cart_id = cart_id;
         this.user_id = user_id;
@@ -32,7 +32,7 @@ public final class Cart {
         return cart_id;
     }
 
-    public void setCartId(long cart_id) {
+    public void setCartId(Integer cart_id) {
         this.cart_id = cart_id;
     }
 
@@ -40,7 +40,7 @@ public final class Cart {
         return user_id;
     }
 
-    public void setUserId(long user_id) {
+    public void setUserId(Integer user_id) {
         this.user_id = user_id;
     }
 
@@ -48,15 +48,15 @@ public final class Cart {
         return product_id;
     }
 
-    public void setProductId(long product_id) {
+    public void setProductId(Integer product_id) {
         this.product_id = product_id;
     }
 
-    public Long getVariantId() {
+    public Integer getVariantId() {
         return variant_id;
     }
 
-    public void setVariantId(Long variant_id) {
+    public void setVariantId(Integer variant_id) {
         this.variant_id = variant_id;
     }
